refactor/ActivityLogger 리랜더링 최적화를 위한 컴포넌트 분리
