+++ src/utils/currentTime.ts
